requirements.txt
asgiref==3.8.1
cctx==1.0.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
dj-rest-auth==6.0.0
Django==5.1
django-allauth==64.0.0
django-encrypted-model-fields==0.6.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.2
